requirements.txt
tensorflow==2.3.0
pillow==6.2.0
numpy==1.19.1
streamlit==0.62.0
